requirements.txt
imageio==2.36.0
matplotlib==3.10.7
numpy==1.21.5
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
opencv_python_headless==4.12.0.88
Pillow==9.0.1
Pillow==12.0.0
rembg==2.0.67
scipy==1.8.0
